+++ djogi/src/jsonb/mod.rs
@@ -45,7 +45,7 @@ pub mod schema;
 pub mod unknown;
 
 pub use mirjzson::{MirJzSON, MirJzSONError};
-pub use path::JsonbPathRef;
+pub use path::{JsonbPathRef, JsonbSqlCast};
 pub use schema::JsonbSchema;
 pub use unknown::{UnknownField, UnknownFieldError, UnknownFieldExt};